requirements.txt
fastapi==0.109.2
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
pydantic==2.6.1
pillow==10.2.0
python-multipart==0.0.7
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
uvicorn==0.27.1